pyproject.toml
[tool.poetry]
name = "folder-organazer"
version = "0.1.0"
description = ""
authors = ["Yuriy Titov <yurtitov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
hydra-core = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
